fix(ds): commonize the modal position
