+++ packages/design-systems/src/components/composite/Datagrid/Density/Density.tsx
@@ -24,6 +24,7 @@ import { HStack } from "@components/patterns/HStack";
 import { Button } from "@components/Button";
 import { Text } from "@components/Text";
 import { LOCALIZATION_EN } from "@locales";
+import { useGetBoxPosition } from "../useGetBoxPosition";
 
 export const Density = <TData extends Record<string, unknown>>(
   props: DensityProps<TData>,
@@ -50,6 +51,8 @@ export const Density = <TData extends Record<string, unknown>>(
       value: "lg",
     },
   ];
+
+  const getBoxPosition = useGetBoxPosition(densityButtonRef);
 
   if (!props.table.enableDensity) {
     return null;
@@ -82,12 +85,12 @@ export const Density = <TData extends Record<string, unknown>>(
         w={"xs"}
         borderRadius={"4px"}
         boxShadow="lg"
-        position={"absolute"}
-        top={"100px"}
         backgroundColor={"bg.default"}
-        zIndex={2}
         ref={densityRef}
         display={densityOpen ? "block" : "none"}
+        style={{
+          ...getBoxPosition(),
+        }}
       >
         <RadioGroup
           options={options}
