+++ packages/design-systems/src/components/composite/Datagrid/SearchFilter/useCustomFilter.ts
@@ -228,19 +228,6 @@ export const useCustomFilter = <TData>({
     [],
   );
 
-  const getBoxPosition = (): CSSProperties => {
-    const box = filterButtonRef.current?.getBoundingClientRect();
-    if (!box) {
-      return {};
-    }
-
-    return {
-      zIndex: 1500, // ark-ui modal has z-index of 1400. So, we need to set it higher than that.
-      position: "absolute",
-      top: Math.ceil(box.bottom - box.top),
-    };
-  };
-
   return {
     filterRef,
     filterButtonRef,
@@ -251,7 +238,6 @@ export const useCustomFilter = <TData>({
     clearFilterHandler,
     addNewFilterRowHandler,
     filterChangedHandler,
-    getBoxPosition,
     generateGraphQLQueryFilter: generateFilter, // For testing purpose
     setPrevFilter, // For testing purpose
   };
